Refactor AddCreditToWallet to use Controller
