+++ Credit-Wallet/Controllers/WalletController.cs
@@ -1,4 +1,5 @@
-﻿using Credit_Wallet.Features.MakeWallet;
+﻿using Credit_Wallet.Features.AddCreditToWallet;
+using Credit_Wallet.Features.MakeWallet;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Credit_Wallet.Controllers;
@@ -8,10 +9,13 @@ namespace Credit_Wallet.Controllers;
 public class WalletController : ControllerBase
 {
     private readonly IMakeWalletService _makeWalletService;
+    private readonly AddCreditToWalletHandler _addCreditToWalletHandler;
 
-    public WalletController(IMakeWalletService makeWalletService)
+    public WalletController(IMakeWalletService makeWalletService,
+                            AddCreditToWalletHandler addCreditToWalletHandler)
     {
         _makeWalletService = makeWalletService;
+        _addCreditToWalletHandler = addCreditToWalletHandler;
     }
 
     [HttpPost("create")]
@@ -20,5 +24,15 @@ public class WalletController : ControllerBase
         var resultId = await _makeWalletService.HandleAsync();
         
         return Ok(new {message = "Wallet created successfully"});
+    }
+    [HttpPost("api/wallet/add-credit")]
+    public async Task<IActionResult> AddCreditAsync([FromBody] AddCreditToWalletRequest request)
+    {
+        var response = await _addCreditToWalletHandler.HandleAsync(request);
+        if (!response.Success)
+        {
+            return BadRequest(response);
+        }
+        return Ok(response);
     }
 }
